Split 'BBBackend' and 'BBBNativeBackend' in categories.
