+++ src/api/BBUser.ts
@@ -25,7 +25,7 @@ export default class BBUser {
                 userId: this.userId
             };
 
-            Backend.getBackend().getEnrolledCourses(parameters).then((information) => {
+            Backend.getBackend().courses.getEnrolledCourses(parameters).then((information) => {
                 this.enrolledCourses = information;
                 resolve(this.enrolledCourses);
             });
