+++ src/backend/BBNativeBackend/BBMisc.ts
@@ -1,0 +1,8 @@
+import {HTTPRequest} from '../../common';
+import Misc from '../../common/BBBackend/misc';
+
+export default class BBMisc extends Misc {
+  public getBlackboardDomain(): string {
+      throw new Error("Method not implemented.");
+  }
+}
